Put list action tests in correct files
