+++ bookwyrm/tests/views/lists/test_list.py
@@ -3,6 +3,7 @@ import json
 from unittest.mock import patch
 
 from django.contrib.auth.models import AnonymousUser
+from django.core.exceptions import PermissionDenied
 from django.template.response import TemplateResponse
 from django.test import TestCase
 from django.test.client import RequestFactory
@@ -13,6 +14,7 @@ from bookwyrm.tests.validate_html import validate_html
 
 
 # pylint: disable=unused-argument
+# pylint: disable=too-many-public-methods
 class ListViews(TestCase):
     """list view"""
 
@@ -235,3 +237,472 @@ class ListViews(TestCase):
         self.assertEqual(self.list.description, "wow")
         self.assertEqual(self.list.privacy, "direct")
         self.assertEqual(self.list.curation, "curated")
+
+    def test_delete_list(self):
+        """delete an entire list"""
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.local_user,
+                book=self.book,
+                approved=True,
+                order=1,
+            )
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.local_user,
+                book=self.book_two,
+                approved=False,
+                order=2,
+            )
+        request = self.factory.post("")
+        request.user = self.local_user
+        with patch(
+            "bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"
+        ) as mock, patch("bookwyrm.lists_stream.remove_list_task.delay") as redis_mock:
+            views.delete_list(request, self.list.id)
+        self.assertTrue(redis_mock.called)
+        activity = json.loads(mock.call_args[1]["args"][1])
+        self.assertEqual(activity["type"], "Delete")
+        self.assertEqual(activity["actor"], self.local_user.remote_id)
+        self.assertEqual(activity["object"]["id"], self.list.remote_id)
+        self.assertEqual(activity["object"]["type"], "BookList")
+
+        self.assertEqual(mock.call_count, 1)
+        self.assertFalse(models.List.objects.exists())
+        self.assertFalse(models.ListItem.objects.exists())
+
+    def test_delete_list_permission_denied(self):
+        """delete an entire list"""
+        request = self.factory.post("")
+        request.user = self.rat
+        with self.assertRaises(PermissionDenied):
+            views.delete_list(request, self.list.id)
+
+    def test_add_book(self):
+        """put a book on a list"""
+        request = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request.user = self.local_user
+
+        with patch(
+            "bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"
+        ) as mock:
+            views.add_book(request)
+            self.assertEqual(mock.call_count, 1)
+            activity = json.loads(mock.call_args[1]["args"][1])
+            self.assertEqual(activity["type"], "Add")
+            self.assertEqual(activity["actor"], self.local_user.remote_id)
+            self.assertEqual(activity["target"], self.list.remote_id)
+
+        item = self.list.listitem_set.get()
+        self.assertEqual(item.book, self.book)
+        self.assertEqual(item.user, self.local_user)
+        self.assertTrue(item.approved)
+
+    def test_add_two_books(self):
+        """
+        Putting two books on the list. The first should have an order value of
+        1 and the second should have an order value of 2.
+        """
+        request_one = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_one.user = self.local_user
+
+        request_two = self.factory.post(
+            "",
+            {
+                "book": self.book_two.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_two.user = self.local_user
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            views.add_book(request_one)
+            views.add_book(request_two)
+
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book)
+        self.assertEqual(items[1].book, self.book_two)
+        self.assertEqual(items[0].order, 1)
+        self.assertEqual(items[1].order, 2)
+
+    def test_add_three_books_and_remove_second(self):
+        """
+        Put three books on a list and then remove the one in the middle. The
+        ordering of the list should adjust to not have a gap.
+        """
+        request_one = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_one.user = self.local_user
+
+        request_two = self.factory.post(
+            "",
+            {
+                "book": self.book_two.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_two.user = self.local_user
+
+        request_three = self.factory.post(
+            "",
+            {
+                "book": self.book_three.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_three.user = self.local_user
+
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            views.add_book(request_one)
+            views.add_book(request_two)
+            views.add_book(request_three)
+
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book)
+        self.assertEqual(items[1].book, self.book_two)
+        self.assertEqual(items[2].book, self.book_three)
+        self.assertEqual(items[0].order, 1)
+        self.assertEqual(items[1].order, 2)
+        self.assertEqual(items[2].order, 3)
+
+        remove_request = self.factory.post("", {"item": items[1].id})
+        remove_request.user = self.local_user
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            views.remove_book(remove_request, self.list.id)
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book)
+        self.assertEqual(items[1].book, self.book_three)
+        self.assertEqual(items[0].order, 1)
+        self.assertEqual(items[1].order, 2)
+
+    def test_adding_book_with_a_pending_book(self):
+        """
+        When a list contains any pending books, the pending books should have
+        be at the end of the list by order. If a book is added while a book is
+        pending, its order should precede the pending books.
+        """
+        request = self.factory.post(
+            "",
+            {
+                "book": self.book_three.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request.user = self.local_user
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.local_user,
+                book=self.book,
+                approved=True,
+                order=1,
+            )
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.rat,
+                book=self.book_two,
+                approved=False,
+                order=2,
+            )
+            views.add_book(request)
+
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book)
+        self.assertEqual(items[0].order, 1)
+        self.assertTrue(items[0].approved)
+
+        self.assertEqual(items[1].book, self.book_three)
+        self.assertEqual(items[1].order, 2)
+        self.assertTrue(items[1].approved)
+
+        self.assertEqual(items[2].book, self.book_two)
+        self.assertEqual(items[2].order, 3)
+        self.assertFalse(items[2].approved)
+
+    def test_approving_one_pending_book_from_multiple(self):
+        """
+        When a list contains any pending books, the pending books should have
+        be at the end of the list by order. If a pending book is approved, then
+        its order should be at the end of the approved books and before the
+        remaining pending books.
+        """
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.local_user,
+                book=self.book,
+                approved=True,
+                order=1,
+            )
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.local_user,
+                book=self.book_two,
+                approved=True,
+                order=2,
+            )
+            models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.rat,
+                book=self.book_three,
+                approved=False,
+                order=3,
+            )
+            to_be_approved = models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.rat,
+                book=self.book_four,
+                approved=False,
+                order=4,
+            )
+
+        view = views.Curate.as_view()
+        request = self.factory.post(
+            "",
+            {
+                "item": to_be_approved.id,
+                "approved": "true",
+            },
+        )
+        request.user = self.local_user
+
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            view(request, self.list.id)
+
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book)
+        self.assertEqual(items[0].order, 1)
+        self.assertTrue(items[0].approved)
+
+        self.assertEqual(items[1].book, self.book_two)
+        self.assertEqual(items[1].order, 2)
+        self.assertTrue(items[1].approved)
+
+        self.assertEqual(items[2].book, self.book_four)
+        self.assertEqual(items[2].order, 3)
+        self.assertTrue(items[2].approved)
+
+        self.assertEqual(items[3].book, self.book_three)
+        self.assertEqual(items[3].order, 4)
+        self.assertFalse(items[3].approved)
+
+    def test_add_three_books_and_move_last_to_first(self):
+        """
+        Put three books on the list and move the last book to the first
+        position.
+        """
+        request_one = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_one.user = self.local_user
+
+        request_two = self.factory.post(
+            "",
+            {
+                "book": self.book_two.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_two.user = self.local_user
+
+        request_three = self.factory.post(
+            "",
+            {
+                "book": self.book_three.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request_three.user = self.local_user
+
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            views.add_book(request_one)
+            views.add_book(request_two)
+            views.add_book(request_three)
+
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book)
+        self.assertEqual(items[1].book, self.book_two)
+        self.assertEqual(items[2].book, self.book_three)
+        self.assertEqual(items[0].order, 1)
+        self.assertEqual(items[1].order, 2)
+        self.assertEqual(items[2].order, 3)
+
+        set_position_request = self.factory.post("", {"position": 1})
+        set_position_request.user = self.local_user
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            views.set_book_position(set_position_request, items[2].id)
+        items = self.list.listitem_set.order_by("order").all()
+        self.assertEqual(items[0].book, self.book_three)
+        self.assertEqual(items[1].book, self.book)
+        self.assertEqual(items[2].book, self.book_two)
+        self.assertEqual(items[0].order, 1)
+        self.assertEqual(items[1].order, 2)
+        self.assertEqual(items[2].order, 3)
+
+    def test_add_book_outsider(self):
+        """put a book on a list"""
+        self.list.curation = "open"
+        self.list.save(broadcast=False)
+        request = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request.user = self.rat
+
+        with patch(
+            "bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"
+        ) as mock:
+            views.add_book(request)
+            self.assertEqual(mock.call_count, 1)
+            activity = json.loads(mock.call_args[1]["args"][1])
+            self.assertEqual(activity["type"], "Add")
+            self.assertEqual(activity["actor"], self.rat.remote_id)
+            self.assertEqual(activity["target"], self.list.remote_id)
+
+        item = self.list.listitem_set.get()
+        self.assertEqual(item.book, self.book)
+        self.assertEqual(item.user, self.rat)
+        self.assertTrue(item.approved)
+
+    def test_add_book_pending(self):
+        """put a book on a list awaiting approval"""
+        self.list.curation = "curated"
+        self.list.save(broadcast=False)
+        request = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request.user = self.rat
+
+        with patch(
+            "bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"
+        ) as mock:
+            views.add_book(request)
+
+        self.assertEqual(mock.call_count, 1)
+        activity = json.loads(mock.call_args[1]["args"][1])
+
+        self.assertEqual(activity["type"], "Add")
+        self.assertEqual(activity["actor"], self.rat.remote_id)
+        self.assertEqual(activity["target"], self.list.remote_id)
+
+        item = self.list.listitem_set.get()
+        self.assertEqual(activity["object"]["id"], item.remote_id)
+
+        self.assertEqual(item.book, self.book)
+        self.assertEqual(item.user, self.rat)
+        self.assertFalse(item.approved)
+
+    def test_add_book_self_curated(self):
+        """put a book on a list automatically approved"""
+        self.list.curation = "curated"
+        self.list.save(broadcast=False)
+        request = self.factory.post(
+            "",
+            {
+                "book": self.book.id,
+                "book_list": self.list.id,
+                "user": self.local_user,
+            },
+        )
+        request.user = self.local_user
+
+        with patch(
+            "bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"
+        ) as mock:
+            views.add_book(request)
+            self.assertEqual(mock.call_count, 1)
+            activity = json.loads(mock.call_args[1]["args"][1])
+            self.assertEqual(activity["type"], "Add")
+            self.assertEqual(activity["actor"], self.local_user.remote_id)
+            self.assertEqual(activity["target"], self.list.remote_id)
+
+        item = self.list.listitem_set.get()
+        self.assertEqual(item.book, self.book)
+        self.assertEqual(item.user, self.local_user)
+        self.assertTrue(item.approved)
+
+    def test_remove_book(self):
+        """take an item off a list"""
+
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            item = models.ListItem.objects.create(
+                book_list=self.list,
+                user=self.local_user,
+                book=self.book,
+                order=1,
+            )
+        self.assertTrue(self.list.listitem_set.exists())
+
+        request = self.factory.post("", {"item": item.id})
+        request.user = self.local_user
+
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            views.remove_book(request, self.list.id)
+        self.assertFalse(self.list.listitem_set.exists())
+
+    def test_remove_book_unauthorized(self):
+        """take an item off a list"""
+        with patch("bookwyrm.models.activitypub_mixin.broadcast_task.apply_async"):
+            item = models.ListItem.objects.create(
+                book_list=self.list, user=self.local_user, book=self.book, order=1
+            )
+        self.assertTrue(self.list.listitem_set.exists())
+        request = self.factory.post("", {"item": item.id})
+        request.user = self.rat
+
+        with self.assertRaises(PermissionDenied):
+            views.remove_book(request, self.list.id)
+        self.assertTrue(self.list.listitem_set.exists())
+
+    def test_save_unsave_list(self):
+        """bookmark a list"""
+        self.assertFalse(self.local_user.saved_lists.exists())
+        request = self.factory.post("")
+        request.user = self.local_user
+        views.save_list(request, self.list.id)
+        self.local_user.refresh_from_db()
+        self.assertEqual(self.local_user.saved_lists.first(), self.list)
+
+        views.unsave_list(request, self.list.id)
+        self.local_user.refresh_from_db()
+        self.assertFalse(self.local_user.saved_lists.exists())
